Reland "Don't install instrumentation stubs for single thread deopts"
This reverts commit 2260840364e5f189edb7909318f4e7841f0ecd9c.

Reason for revert: This is a reland after fixing the failures. The
fixes are:
1. When Updating methods code we can use fast path only when there is
no instrumentation including instrumenting the stack.
2. When restoring the instrumentation stack we could find some
runtime methods with instrumentation exit pc when running quick to
interpreter bridge. These should be included in the frames_removed
count.
3. Also set the jit threshold for jvmti to 1. JitOnFirstUse doesn't play
well with redefinition since the jit compilations happen on regular
threads and it is not easy to stop them when redefining.


Change-Id: I50f4cc6903310b62ca1c49050b6cd7320feca89e