Don't enqueue new compilations when redefinition is in progress
When redefining a method we stop any new compilations by stopping the
jit thread pool but we used to allow enqueuing new compilations. Since
redefinition could obsolete some of these methods it is better we stop
enqueuing any new compilations. This also prevents any unintended
replacements of the method's declaring class when replacing the old
classes.

This CL adds a new is_suspended_ which is set in the ScopedJitSuspend.
When we are in this scope we don't allow any new compilations to
progress.

Bug: 234888286
Bug: 206029744
Test: art/test.py
Change-Id: I1c6eea80f06519ea2cbdfbd8777f9d7de568fd62