Reland "Don't use instrumentation stubs for native methods in debuggable"
This reverts commit 5c9b55aa95295a287abd86f1e7fbe98c3f35ffd6.

Reason for revert: Relanding with fixes for failure
Fixes:
1. Arm64 needs to use 64-bit registers
2. We cannot deoptimize directly from GenericJniEndTrampoline. Instead
set an exception and handle the exception later after returning to
art_quick_generic_jni_trampoline where the context is properly setup.
To indicate that the exception was caused by MethodExitEvent, introduce
a new MethodExitException. The original exception should be stored on
the exception context.

Change-Id: If1f08eca69626f60f42f10205b482a3764610846