Fix bugprone-unused-return-value clang-tidy issues
There is only one place where this check was failing, and this is
in CollectionVector::SortByMapOrder in dexlayout/dex_ir.h - there
the code was calling std::unique_ptr::release() to prevent a case
where a double free could occur. This CL changes the way in which
the sorting is performed so that no pointer release is required.

Bug: 213953102
Test: m tidy-art
Test: atest art_dexlayout_tests (especially DexLayoutTest.DexFileOutput)
Change-Id: I525299c347d89171f91e10323d5aa8702e22b8d5