Move existing library loading tests to a TF hostside test.
Retire the manual runtest.sh script with something that tradefed can
run. It requires root and will remount, push .so libraries and
public.libraries*txt files to any path they need to go to, and then do
a soft reboot.

Since libraries are pushed to the test locations rather than being
installed through LOCAL_MODULE_PATH, the shared libraries in the build
don't matter, and the same one is now used for all locations. That
means it no longer exercises the `product_specific` Soong property for
the two product libraries.

Test: atest libnativeloader_e2e_tests
Bug: 137356719
Change-Id: Ida1c70cb282946d72736da5cab0c751b60f0b03a