[Sim] Make dlsymLookup trampoline native.
Compilers's trampolines are emitted to oat files
to act as extenders for the calls from the compiled
code - when the calls target is further away than
what the immediate offset of the call can encode (this
also reduces code size).

In simulator mode the dlsymLookup trampoline is executed
in native context and thus should be compiled as native.

Test: test-art-simulator.
Test: 660-clinit.

Change-Id: I03b385dcf74568d00a283a2ee5d9198e4a2010ff