Don't validate boot image oat files when validating app oat header.
aosp/1915157 added a check on oat files of boot images. It was meant to
be only executed when the runtime loads the oat files. However, the code
is unexpectedly executed in the slow path of validating app oat header.
After the ART module is updated (through adb install or Play), it is
expected that the slow path is taken when ART is validating the
out-dated app oat files on /system, so we should avoid validating boot
image oat files in the slow path because it regresses app startup time.

Bug: 238394854
Test: -
  1. adb install the ART module.
  2. Open an app that has an out-dated oat file on /system.
  3. Capture the trace.
  4. No longer see "Open oat file ..." for boot image oat files in the
     trace.
Change-Id: I2b16b68e739b6659c33b3e4d1237df97fe51880a