Revert "Propagating values from if clauses to its successors"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/98548ca7-e251-4500-8697-859f2aa5ed08). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2d968f23274aede173023bb224ff82390d025a17