ART: Query change in behaviour caused by 4f18b94
NOT FOR MERGING!

The purpose of this CL is to inquire regarding the specific behaviour of
"Make linear loop optimization safe from overflow (4f18b94).

Between patch-set 3 and patch-set 5, the logic to avoid narrowing linear
induction and type mismatch when generating last-value replacement for
linear induction variables was inverted in order to return early instead
of guarding the entire method body. However, instead of an exact logical
inversion the conditional check was altered slightly, and this new check
in PS5 produces some different behaviour (see below). We wanted to
enquire whether this behaviour is expected?

For reference:

(Cond1) the original conditional check in PS3:
if (!HInductionVarAnalysis::IsNarrowingLinear(info) && trip->type ==
info->type)

(Cond2) An exact inversion would be:
if (HInductionVarAnalysis::IsNarrowingLinear(info) || trip->type !=
info->type)

(Cond3) The new/current conditional check in PS5:
if (HInductionVarAnalysis::IsNarrowingLinear(info) && trip->type ==
info->type)

This change includes two checker-tests which highlight the change in
behaviour between PS3 and PS5:

In the case of closedInductionUnmatchedTypesNotOptimized:
- The Phi which would be replaced by last value at its use outside of
the loop has type Int64, and so do its op_a and op_b IVs - therefore
this is not a narrowing linear IV (IsNarrowingLinear(info) == false)
- The trip-count has type Int32. This means that trip->type ==
info->type evaluates to false
- Therefore neither condition is met and we do not return early in
patch-set 5 and the last value replacement occurs; however in both the
base code and patch-set 3 we do return early and do not generate the
last value replacement

In the case of narrowingLinearInductionNotOptimized:
- The Phi which would be replaced by last value at its use outside of
the loop has type Int16, but its op_a and op_b IVs both have type Int32
- therefore this is a narrowing linear IV (IsNarrowingLinear(info) ==
true)
- The trip-count has type Int32. This means that trip->type ==
info->type evaluates to false
- Therefore only the first condition is met and we do not return early
in patch-set 5 and the last value replacement occurs; however in both
the base code and patch-set 3 we do return early and do not generate the
last value replacement

A Note: as far as I can tell - GenerateLastValueLinear is always
called with a Loop Header Phi, and then this calls GenerateCode to
handle the rest of the factored-use-def (FUD) chain. Whilst other IVs in
the chain may be narrowing as a result of explicit casts, the only
situations in which a Phi is narrowing is when its type is smaller than
Int32 and the inputs are implicitly narrowing; however the type of a
TripCount induction variable will never be less than Int32. Provided I
am correct in this assessment, in GenerateLastValueLinear
`IsNarrowingLinear(info)` will be true iff `trip->type == info->type` is
false, so the test in PS5 (Cond3) can never evaluate to true.

Test: 618-checker-induction

Change-Id: I0ccdc21dc6991ed949e83a5471a31f8e0aebef2b