Revert "Explicitly madvise boot image oat files"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/94de7f2c-ed30-4358-9083-eecc0bb10644). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3c02852ac8ada2a1f7d4ae9985746327c69e7fae