Set a continuous reporting spec for MetricsReporterTest#SampleRate* tests
The tests SampleRateEnable50 and SampleRateEnableAll have sometimes been
flaky because of the ShouldContinueReporting() assertion being false
when it was supposed to be true. Looking at the boolean conditions that
compose the return value of that method, one that could turn out to be false is the one related to the report interval index being still within
the period spec size (e.g., if the test for any reason ends up taking
more than one second before reaching that assertion, it could be the
case that the background thread has reported metrics and moved the index
outside the period spec). Considering these tests are exercising the
logic around the session_id and the reporting_mods values (and not the
reporting period itself), changing the period spec to be repeating
should remove one possible factor of flakiness for these tests. Also,
this commit fixes a few typos and incorrect statements in comments.

Bug: 198418961
Test: atest art_runtime_tests
Change-Id: I30cc0bccbd5e144542672b3deba6aca21d9f855c