Figure out the right lib(64) path in a more principled way.
The old approach would copy the TARGET_ARCH libraries to both 32 and 64
bit directories if both exist in the TARGET_OUT tree. That may not lead
to observable bugs because only the directory corresponding to
TARGET_ARCH will be actually used later, but that's not a safe
assumption, and it's still confusing if nothing else.

Test: lunch armv8-eng; m build-art-target-golem
      lunch arm_krait-eng; m build-art-target-golem
  Check contents of $(TARGET_OUT/{lib,lib64} directories after both
  builds
Bug: 243935575
Change-Id: I763644feccea7f95b4ea7bde87ac3e048178a51d