Remove TryRemovingNullCheck from LSE
TryRemovingNullCheck would remove null checks right before eliminated loads. However, it was doing it without really checking if the object it was loading from was null or not.

This has been there since the first iteration of LSE, and my guess why it was implemented it was to remove NullChecks with no uses (i.e. it was assuming that the NullCheck was only uses by the load which was being eliminated).

Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Change-Id: I96be9650f8b16fee7fcca114b9b330d4f3b7e0b4