Revert "Remove RBE toolchain_inputs bug work-around"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/db23fc33-d25c-4be9-9da1-da91bb0ec27c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iddafe216e93b38756f29cbfb25358cc634295ecd