Add tests that try to load non-shared native libs in each partition.
One test case in each app that tries to load them directly, and another
one that goes through a shared library in the system partition.

Also set the appropriate Soong properties on the test apps, for what it
may be worth (doesn't seem to have any effect that we don't get by just
pushing them to the various locations).

Test: atest libnativeloader_e2e_tests
Bug: 137356719
Bug: 237577392
Change-Id: I9e2c8162e0f0928020355a507b90f0fa76bf6b6b