Revert "Remove TryRemovingNullCheck from LSE"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/9bbf8721-3fac-482d-b80d-3373817f8f47). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I15836b77c5b4d2643141c23470748767f2f2c826