Check if method is invokable when initializing method's code
When initializing a method, we need to check if the method is invokable
and if it isn't invokable we should use switch interpreter. We could see
non-invokable methods in some cases for ex: when deoptimizing a boot
image or when initializing the code after verification. Once the class
is verified, we update the entry points of method that have interpreter
entry points to use Nterp where possible. However, if the method isn't
invokable we should retain an interpreter entry point.

Bug: 206029744
Test: art/testrunner.py --trace
Change-Id: Idf2e95e29791d0e4244afa5dfe36e5b78649853c