Check if method is invokable when initializing method's code
When initializing a method, we need to check if the method is invokable
and if it isn't invokable we should use switch interpreter. We could see
non-invokable methods in some cases for ex: when deoptimizing a boot
image or when initializing the code after verification. When updating
the entrypoints for instrumentation, we need to take care not to use
instrumentation stubs for such methods.

Bug: 206029744
Test: art/testrunner.py -t 2240-tracing-non-invokable-method
Change-Id: Idf2e95e29791d0e4244afa5dfe36e5b78649853c