Allow the vendor partition to be in /system/vendor.
It's uncertain whether this happens in practice, but with the changed
logic in https://r.android.com/2211602 it would otherwise be treated as
a system APK and given full access to system libraries, which would be
unfortunate.

A note on tests: It's difficult to set up fixtures to test
/system/{product,vendor} paths if the device has them as symlinks to
/{product,vendor}, because it would require moving the mount points and
symlink in the other direction.

Test: atest libnativeloader_e2e_tests
Bug: 237577392
Change-Id: I9267c008db644ebe38f6215d3b692b98dd88ee34