Support mixed mode autovectorization.
This CL brings support for mixed mode vectorization -
when a graph could have some loops vectorized in predicated
mode and some others - in traditional. The patch mainly changes
two areas: vectorizer itself and the codegen (mostly arm64).

Vectorizer will try to vectorize each loop in traditional mode
(e.g. using arm64 NEON); if it fails to do so and if the target
supports predicated vectorization mode (e.g. arm64 target with
SVE), the vectorizer will attempt to vectorize the loop in
predicated mode. This heuristic of which mode to use, could be
extended for each target in the future if needed.

Arm64 codegen has been reworked: instead of having two separate
codegen instances for SVE and NEON targets we now have a single
one with helpers for both vector extensions. Which helper/visitor
to used is determined the following way:
 - for ParallelMove helpers this is graph-level decision: if the
   graph has predicated instructions (SVE), than for all parallel
   moves the SVE versions will be used as a conservative measure.
 - vector operations visitors type will be chosen based on whether
   the instruciton is predicated.

Test: test-art-host, test-art-target.
Test: test-art-target for arm64 SVE target.

Change-Id: I48dccac895aafce14fb607f4937454927e75656c