Refactor vectorization general pipeline.
Currently, vectorization methods such as GenerateNewLoop and
Vectorize process loops for both traditional and predicated
modes with multiple IsInPredicatedVectorizationMode() checks
which control the behavior. This makes the code have many
implicit assumptions and rather difficult to extend.

The aim of this CL is to create two separate pipelines
for traditional and predicated vectorization and to share
as much common helpers as possible.

Original author: Artem Serov <Artem.Serov@linaro.org>

Test: ./art/test/testrunner/testrunner.py --host --optimizing --jit
Test: ./art/test/testrunner/testrunner.py --target --optimizing --jit

Test: target tests on arm64 with SVE (for details see
      art/test/README.arm_fvp).

Change-Id: I35a57e8585aa0d24aaf7cdcdd852199e24767172