Fix .so library separation between the ART APEX and system.
The previous approach that added /apex/com.android.art/lib(64) to
LD_LIBRARY_PATH mixes up ART APEX and system libs in the same linker
namespace and doesn't work if they depend on different versions of
libraries with unstable ABIs (e.g. libbase.so).

This CL fixes it by adapting
/linkerconfig/com.android.art/ld.config.txt for the tests and feeding
that to the linker with LD_CONFIG_FILE. That way the test binaries get
loaded in the same linker namespace as the real APEX binaries and can
access all internal libraries correctly.

The problem with this approach is that LD_CONFIG_FILE is supported only
in debug builds, so test configs for CTS tests are left out for now.

Test: Start an AVD with a platform image from tm-release (ab/9012097),
      adb install an ART module from head, and run:
  atest art_standalone_compiler_tests art_standalone_dex2oat_tests \
    art_standalone_dexoptanalyzer_tests \
    art_standalone_libdexfile_tests art_standalone_libprofile_tests \
    art_standalone_oatdump_tests art_standalone_profman_tests \
    art_standalone_runtime_compiler_tests \
    art_standalone_runtime_tests \
    art_standalone_libdexfile_external_tests \
    art_standalone_libdexfile_support_tests \
    art_standalone_libartbase_tests art_standalone_dexdump_tests \
    art_standalone_dexlist_tests art_standalone_dexpreopt_tests
  (art_standalone_dexpreopt_tests failed, but it wasn't a regression)
Bug: 247108425
Change-Id: Ic9710793f5421c4ee569826c3bf47b425536b649