Revert "Regenerate ART test files (2022-09-08)."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/dcd78646-9f62-48d8-9499-a6193744abf7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4eaa308db603eb6d4922a8acd81d39e74098c711