Add code and rules to build golemtiagent.so to use in golem benchmarks
We are updating debuggable-cc configuration in golem benchmarks to
measure the performance of debugger. As a first step, we are adding
a breakpoint to measure the overhead of instrumentation related code.
To achieve this we need a tiagent to be able to set a breakpoint.
This CL adds a minimal agent that adds OnLoad callback and necessary
methods to set a breakpoint.

Bug: 232212577
Test: m build-art-host-golem
Change-Id: I2da3770ff85edb46da64d09ec0047ec6e7564fd0