run-test-jar: Adjust test verbosity
Add --verbose flag (and remove the --dev and --quiet flags).

If the verbose flag is enabled, unconditionally print all executed
commands, their output and other logging messages.

If the verbose flag is disabled, disable all output, except for
the output of dalvikvm command which is needed to compare to the
actual stdout and stderr to the expected files.

Test: ./art/test.py -r -t 001
Change-Id: Idfdf3ae46581894c6257d93fee5eb7933d09ecf3