Revert "Cleanup libart-compiler version script"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/17fb56cb-a86b-414b-b914-448b34621d03). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If193822898963149a3ae0b6233f3c541a95b0da3