Revert "Cleanup libart-compiler version script"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/62fec752-9613-4b0a-93dc-3c978b3338bc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6433c7c10245b89f80a9fcc50a2983b8a2809051