Revert "Cleanup libart-compiler version script"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/cf5cd1ba-b115-4fb3-b78d-19c23587d09c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If859854dbfe5cf101500ced064f3d9cb3117849f