Revert "Cleanup libart-compiler version script"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/3d296ccb-0a7e-4c17-b3ac-adaaae59161c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia13c22ac443057b365c0049c3136c07be24aa5fe