Revert "Cleanup libart-compiler version script"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/6280672a-67ad-4bad-bd07-bb36195ad879). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If9153fb601cba1eeeacf10e9ecb8e3236b244bb7