Create framework-art.
This is a java_sdk_library that will be used for exposing APIs to apps.

The list of ART bootclasspath jars is hardcoded in more than 10 places
across multiple projects. To avoid updating the hardcoded list, this
library is statically linked into "core-libart" instead of being a
standalone jar in bootclasspath.

Bug: 248580326
Test: m
Change-Id: I8ca01de50ef9dcd76e034bf115a68ddf92b2c59e