Revert "Fix panics in art build code when target arch is riscv64"
Revert "Use MinApiForArch to set minSdkVersion for apexes"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/7bb39570-3729-4a1c-a59e-56b603604d24). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iecd407e177630cf3d80d550c00a4ca4ffc8b29ba