Revert "Move `VerificationResults` to dex2oat/."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/4486e97c-a204-4743-9ddc-1f18d9056133). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3ec6afce1c0386537337d7aee3b303d873c1f72a