Revert "ART: Move `CompiledMethod{,-Storage}` to dex2oat/."
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/1a6e9616-24a3-414f-92f1-e8c68f8ff0bd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I84c4370d8a933fc10fc78529ca492c01c9b3a92f