Don't handle deoptimized funciton in method entry hooks
We always update entrypoint of the method to point to InterpreterBridge
if the method is deoptimized. So, we shouldn't be seeing deoptimized
methods executing artMethodEntryHook which is only called from JITed
code. It is possible that the method was deoptimized after invoking
the method at the first suspend point but those should be handled when
returning from the suspend points.

Bug: 253232638
Test: art/test.py
Change-Id: Iadb25a96307646ad7119162d56038729930767cb