Fix --dump-cfg argument handling in run-test
This patch fixes a bug which was introduced in
d5f69e241e332707e944b92cb68bf76c82781274 when converting the run-test
script from bash to Python. The argument's value is now read as the
user-supplied string instead of a constant string literal "{arg}",
resulting in a correctly produced .cfg file.

Test: test-art-host, test-art-target.

Change-Id: I76e550edcfc66c9746e77d767e830412e0cf90cd