Suspend around Make...VisiblyInitialized
This removes the workarounds introduced in aosp/2212326, and
replaces them with actual fixes. This is easier to do than I had
initially thought. There do not appear to appear to be additional
pointers that could be invalidated by thread suspension.

After this, we no longer disable ithread state checking in any
MakeInitializedClassesVisiblyInitialized calls.

Test: Build and boot AOSP.
Bug: 253691761
Change-Id: I4449c8715e154db504eb94d2da631428bfb52de4