Allow LSE to remove acquire/release operations
If we are synchronizing on a removable singleton, we can guarantee that
no other thread has access to it. We can remove MonitorOperations
related to that singleton since they are redundant. Volatile load/stores
can be treated as regular load/stores.

In the case of MonitorOperation, we can also remove the catch that we
create to catch and rethrow exceptions.

Locally speed compiling we see improvement in code size:
 * System server -0.34%
 * SystemUI: 0.74%
 * AGSA: 0.55%

Bug: 227283233
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Change-Id: Iad51f6576e8d9287824ecbb1220600bf44481f1a