Apply standalone_debug to TypeList and managed classes
Since change [1] in clang compiler, debug information for some types
are missing. Strictly speaking it was not missing for all MANAGED
types, but the MANAGED_PACKED macro was a convenient place to add the
attribute, and it does no damage to the places where it wasn't needed.
This attribute should ensure that debug information is available as it
was before the change [1]. But there might be other types with missing
debug information, but not yet observed.

[1] https://reviews.llvm.org/D106084

Bug: 257296021
Test: llvm-dwarfdump -c --name="TypeList" --debug-info libdexfile.so
Test: llvm-dwarfdump -c --name="DexCache" --debug-info libart.so
Change-Id: Ife70ebf422a65f1967bb44227df2b0c62bc845de