Revert "Fix run-test check for --jvmti-redefine-stress"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/c0d72f87-7651-40f3-8faa-2f4ebad8d298). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idb1454f79eba97ccaea345304c84ffdb1bf0555e