Revert "Fix run-test check for --jvmti-redefine-stress"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/57f3b311-7cf6-4b14-b8f7-953419df7a94). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I49978b2c83940e881df2701b77573ce188c449b6