[WIP] Fix trip count overflow in induction variable opt
NOT FOR SUBMISSION.

The last-value-replacement optimization for periodic
induction variables contains a bug where code can be generated for
trip-count expressions that may overflow and result in incorrect
post-optimization results. This is fixed by only generating code if we
can be certain that the trip count expression is overflow safe.

This also fixes current BCE targeting as dynamic BCE will be
incorrectly performed on loops that have trip counts known at compile
time if the trip count expression will overflow 32-bit integer
arithmetic.

Test: art_compiler_test, 618-checker-induction, 654-checker-periodic
Change-Id: I3f9850ba5b208dd07be3dea1cbc5f4dc3334d41e