Canonicalize methods when in jni::EncodeArtMethod
jmethodIDs returned by jni::EncodeArtMethod are used by external
interfaces (for ex: jvmti) to match methods. As an example, when single
stepping jvmti returns method id corresponding to the methods on the
stack. To map these method ids to method name / signature the method ids
are matched against the method ids returned by the GetDeclaredMethods of
the class. This generally works but breaks for copied methods. So we
should canonicalize methods before returning the method id. We already
do this when using opaque jni ids.

Bug: 244683447
Test: art/testrunner.py -t step-default-method
Change-Id: Ie2e9225fb8ebf5235f08345a079eb73f28f33961