Make periodic loop optimization safe from overflow
Similar to aosp/2162257 but for periodic loops.

The trip number is calculated by doing:
    `(end - start) + (step - 1) / step`
(Note that we add `step - 1` as a way of doing `ceiling`).
This way of calculating can overflow and produce the wrong result
if end and start are in opposite sides of the spectrum.

We can force `a` to be a constant to guarantee that the right result
will be generated when doing loop optimization.

Bug: 256858062
Fixes: 256858062
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Change-Id: I0ab23ae078ca3a95712b9e839d13888508c9dffe