Revert "Merge libart and libart-compiler into a single library"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/aab0e329-2a9c-4047-8a1d-8e02d93b6c98). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib511df9c7b7575f09fce4abd52e40fe84591bc9f