Revert "Merge libart and libart-compiler into a single library"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/390a2eb5-d091-4470-8634-115e9f8446d2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I07a29b7f2500310d41239ea5a1175b1210cbffce