Revert "Merge libart and libart-compiler into a single library"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/0af0dd16-7d2b-4a0e-b67c-bc1dfd61afeb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9c72fef92cb723e2eb14aba25c2802b166c647e8