Don't assume the class has never been looked at in FastVerify.
When we compile dex files individually and one FastVerify fails, we may
verify more classes that in the one dex file.

Also fix a bug when in VerifierDeps when verifying after failing fast
verification: we need to keep the stored assignable types in case they
come from an existing passes vdex file.

Bug: 257335641
Test: 845-fast-verify
Change-Id: I92f64daff1653ce3d58cee81ad3b461fa152cb84