Add clang-format preupload check of CLs.
This will check that new and changed lines conform to the style defined
in .clang-format. Specific code sections can be excluded if it gets in
the way (some examples included).

Also fix the java style to not line break in "/// CHECK" comments.

Java files in test/ have a mix of different indentation depths, so no
attempt is done to fit the style there.

openjdkjvmti/include/jvmti.h is an imported file that should be ignored
altogether. However, it is not straightforward to configure a whole
file exclude. Upload it with --no-verify if it needs to be updated.

Test: clang-format -i art/dex2oat/dex2oat_options.cc
      clang-format -i art/test/569-checker-pattern-replacement/src/Main.java
  Check that reformatted files look reasonable
Bug: 181877164
Change-Id: I991ff032694bf5063e8516fdf71a95d0b047ffea