Lazily execute the frame size check for nterp.
Looking at the frame size for a method requires pulling in the code
item, which has a significant cost (1% of total cpu cycles during app
startup).

Given doing the check dynamically is cheap and that nterp is only used
until a method is hot, it seems the right tradeoff to move the check
lazily.

Test: test.py
Change-Id: If4a1f71275e103a1225de8053c14f28cb49fd51b