Change the default clock source used by method tracing on linux
We can either use thread cpu / wall clock / both for getting timestamps
for method entry / exit events. The default source for linux is set to
both on linux and to wall clock on others. There is no reason to special
case linux. The thread cpu timestamps are quite slow  when we use then as
a timestamp for each method entry / exit event. Hence setting wall clock
timestamps as default is better for linux too.

Bug: 259258187
Test: art/test.py
Change-Id: I14edee67e6a6fedf41dbc79f32650f817fc64ec5