Revert "Run test run: Filter interleaved log lines."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9cdddf48-6cb0-4546-8233-c3de1f7f3fdd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icf8601f412a9dbfaed9ff7a7c5e4f49b0204336d