Revert "Run test run: Filter interleaved log lines."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/81ff9b30-fb24-4e08-ba94-c35005fabd03). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If0b713355da767afcdea42d011f709b2068002a0