Revert "Add tests for loading native libraries in various locati..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ad9ed3bf-1517-4914-ba83-2a26aedd993e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6351bb31e739897a6fdca92da8f0726023fb992d