Revert "Add tests for loading native libraries in various locati..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/257336fd-f53a-43fe-b079-85d71b0b32fb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2c473b0a81637269ddae0a6c5bd99bca87cb4fe0