Revert "Remove tries which don't contain throwing instructions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/167632cd-62c7-4690-96db-3e496583a1b0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1f79def643580c325c773aa70879009fcc76606c