Keep sigchain out of the crash stack
When a java process runs into a segfault, sigchain handles it and then
aborts if it doesn't actually have a handler. This has the effect of
removing the actual stack of the segfault from any crash dump, since the
signal is handled on an alternate stack with sigaltstack, requiring
shenanigans with gdb and a core dump and ucontexts to find it again.
Instead, let's remove sigchain as a handler and return. This will
make the same segfault happen again, but this time it won't be handled
and the correct stack will appear in the crash dump or debugger.

Change-Id: I04cec8b371dbed20929a930af8c1f1a568eed3ce