Rewrite `LocalReferenceTable`.
Change `jobject` for local references to a `GcRoot<>*` with
a `kLocal` tag, so that decoding can be done by clearing the
tag and dereferencing. Optimize operations with CheckJNI
disabled and small local reference table.

Results for the timeGetBytesAscii#EMPTY benchmark from the
libcore's StringToBytesBenchmark suite on blueline-userdebug
with the cpu frequencies fixed at 1420800 (cpus 0-3; little)
and 1459200 (cpus 4-7; big):
  32-bit little: ~450 -> ~415
  64-bit little: ~455 -> ~415
  32-bit big:    ~200 -> ~180
  64-bit big:    ~200 -> ~180

This CL also opens an opportunity for inlining a fast-path
for local reference decoding to compiled JNI stubs for
methods that return a reference.

Test: m test-art-host-gtest
Test: testrunner.py --host
Bug: 172332525
Change-Id: Id6f2955419b966fd1ad860bace9c23947080f8ed