Revert "Run test: Execute scripts via runner bash script."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/343f71e8-a2b5-4576-9c75-d39750bfb673). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iab4e8b642e39bff35f1871865101a028048146cd