Revert "Update non-moving space references in reverse"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0cacc901-f105-44d3-993a-db19fa3eefee). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id6431a2d774dcbbd544daa3a829ad0dcc99693fe