Revert "Handle userfaultfd API ioctl on older kernel hosts"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6c7959e5-58f1-4fa2-814f-bba0c5ce5095). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I532fe83abb9bcb979b0d499f34a69f2adc6fcfe1